AI/requirements.txt
pymysql
openai
python-dotenv
mysql-connector-python
playsound
sounddevice